Api/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.25
sqlmodel>=0.0.16
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
requests>=2.31.0

# AI Agent Dependencies
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.20
openai>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.0.0
